+++ src/main/java/open/commons/database/AbstractDao.java
@@ -38,8 +38,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ConcurrentSkipListMap;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import open.commons.function.SQLBiFunction;
 import open.commons.utils.IOUtils;
@@ -48,6 +48,16 @@ import open.commons.utils.SQLUtils;
 /**
  * JDBC Connector를 이용하여 SQL을 직접 적용하여 DBMS와 연동하는 클래스.
  * 
+ * <br>
+ * 
+ * <pre>
+ * [개정이력]
+ *      날짜      | 작성자   |   내용
+ * ------------------------------------------
+ * xxxx.xx.xx       xxx         최초작성
+ * 2019. 10. 17.        박준홍         Logger 교체. org.apache.logging.log4j.Logger -> org.slf4j.Logger 로 교체
+ * </pre>
+ * 
  * @since 2018. 5. 23.
  * @author Park_Jun_Hong_(fafanmama_at_naver_com)
  */
@@ -55,7 +65,7 @@ public abstract class AbstractDao implements AutoCloseable {
 
     private final ConcurrentSkipListMap<String, SQLBiFunction<ResultSet, Integer, ?>> CREATORS = new ConcurrentSkipListMap<>();
 
-    protected Logger logger = LogManager.getLogger(getClass());
+    protected Logger logger = LoggerFactory.getLogger(getClass());
 
     private final String driver;
     private final String url;
@@ -123,8 +133,7 @@ public abstract class AbstractDao implements AutoCloseable {
                 this.con.close();
             }
         } catch (SQLException ignored) {
-
-            logger.error(ignored);
+            logger.error("[ignored]", ignored);
 
             throw ignored;
         }
@@ -564,7 +573,7 @@ public abstract class AbstractDao implements AutoCloseable {
                 this.con.rollback();
             }
         } catch (SQLException ignored) {
-            logger.error(ignored);
+            logger.error("[rollback - fail]", ignored);
         }
     }