+++ src/main/java/open/commons/lang/AbstractCloseable.java
@@ -31,18 +31,28 @@ import java.util.HashSet;
 
 import javax.annotation.Resource;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * 자원을 자동으로 해제하는 클래스.
+ * 
+ * <br>
+ * 
+ * <pre>
+ * [개정이력]
+ *      날짜      | 작성자   |   내용
+ * ------------------------------------------
+ * xxxx.xx.xx       xxx         최초작성
+ * 2019. 10. 17.        박준홍         Logger 교체. org.apache.logging.log4j.Logger -> org.slf4j.Logger 로 교체
+ * </pre>
  * 
  * @since 2019. 2. 19.
  * @author Park_Jun_Hong_(fafanmama_at_naver_com)
  */
 public class AbstractCloseable implements AutoCloseable {
 
-    protected Logger logger = LogManager.getLogger(getClass());
+    protected Logger logger = LoggerFactory.getLogger(getClass());
 
     /**
      * 
@@ -86,7 +96,7 @@ public class AbstractCloseable implements AutoCloseable {
                 c.close();
             }
         } catch (Exception ignored) {
-            logger.warn(ignored);
+            logger.warn("ignored", ignored);
         }
     }
 }